backend/requirements.txt
anyio==3.7.0
autopep8 
bcrypt==4.0.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==41.0.1
decorator==5.1.1
defusedxml==0.7.1
dnspython==2.3.0
ecdsa==0.18.0
email-validator==2.0.0.post2
exceptiongroup==1.1.1
fastapi==0.97.0
future==0.18.3
geocoder==1.38.1
geographiclib==2.0
geopy==2.3.0
googlemaps==4.10.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.2
httptools==0.5.0
httpx==0.24.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
oauthlib==3.2.2
orjson==3.9.1
passlib==1.7.4
psycopg2-binary==2.9.6
pyasn1==0.5.0
pycodestyle 
pycparser==2.21
pydantic==1.10.9
PyJWT==2.7.0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
python-social-auth==0.3.6
python3-openid==3.2.0
PyYAML==6.0
ratelim==0.1.6
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
secure-smtplib==0.1.1
sniffio==1.3.0
social-auth-core==4.4.2
SQLAlchemy==1.4.41
sqlalchemy2-stubs==0.0.2a34
sqlmodel==0.0.8
starlette==0.27.0
toml
typing_extensions==4.6.3
ujson==5.8.0
urllib3==2.0.3
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
